Naive n-body: only 'throw' Error, not strings
Error objects have info on their call stack, source code location, etc. We should always throw an instance of the Error class, or some subclass of it, rather than the strings that we were throwing previously. The error object takes an optional string message in its constructor, so we can still add user-defined information to it.