naive: render points as 5px red squares
We replace the previous vertex shader, which derived the point size from the points Z position, with one which always sets the point size to 5.

We also replace the fragment shader, which drew the point as a circle, with faded edges, with one which just draws every pixel as with the full-opacity point color. Because GL considers points to be squares (with a size set by the vertex shader, 5 in the case,) this renders points as a 5px red square.