naive: enable depth testing in renderer
This will allow us to render the edges behind the points, when we do that later.

We set the depth function to LEQUAL (which passes if the incoming depth is less than or equal to the existing value in the depth buffer.) The default, LESS, causes our points to not render since they have an equal depth to the clear depth bit.