naive: make indentation uniform; trim trailing whitespace
The files now use uniform tab indentation. (We can switch to spaces at some point in the future, if we decide, but either way, we should decide on a uniform standard.)

I've also gone through and deleted trailing whitespace from lines with non-whitespace characters. So lines of pure whitespace were untouched (since it's often handy for making text later added to that line properly indented,) but some non-whitespace character with extra spaces or tabs at the end of the line were fixed.