Naive: remove setPointsImmediate; allow setPoints to do the same thing
setPointsImmediate() differed from setPoints in that it didn't convert its passed-in points to a Float32 array first (presuming they already were so.) I changed setPoints() so that it only tries to convert its passed-in points to a Float32 if it detects they aren't already so.

Since setPoints now works for both types of arguments, I removed setPointsImmediate() and changed the calls to it to call setPoints().