Naive: add in support for specifying the dimensions of the graph space
Previously, we hard-coded the graph space to be from (0,0) to (1,1). Those were the dimensions the GL renderer would render on screen (in world coordinates,) and those were the dimensions that simulator would bound the points to.

Now you can specify the size of the graph as an argument to the Nbody constructor. Note that this doesn't change the size of the rendered graph on screen (in terms of pixels); it only changes the size in terms of internal units that the sim and renderer use for computations.