Naive: pass in random values to CL kernels to resolve overlapping points
Previously, there was a problem with overlapping points: because the distance between the points was (effectively) 0, the forces being applied had no direction. Thus, the forces between overlapping points were 0. The problem was that, because all the other forces acting on both points were exactly the same, the points would always move exactly the same way. This meant that these points were effectively 'stuck' together.

This commit changes that. During graph initialization, we generate some (prime) number of random float2's. We then pass these into the CL kernel. When two points overlap, the kernel will use one of these random values to act on the node, instead of the overlapping point. In this way, overlapping points will be dispersed in a random direction.

(CL kernels have no easy way to generate random values themselves, hence the need to pass in some values like this.)