Naive: maybe make latest changes work on old CL? LEO: PLEASE VERIFY
I added a kernel argument, so I needed to set those in SimCL.js. On old WebCL platforms, we also need to set the types of those arguments. I tried to do that, but my new WebCL platform doesn't even have the type enums, so I had to guess. Maybe it will work, maybe not. Please check and verify!