Naive: change gl.flush() to gl.finish() for benchmarking
The details on the difference between the two are a bit murky, but the take-away is that finish() blocks until all GL calls finish, while flush() just says the outstanding calls will finish in finite time.

We want to block until rendering is full finished, at least for now while we benchmark, so we can figure out the real amount of time it's taking to render things.