Naive: in createPoints(), generate n random points (n=16,384 ATM)
Our old setPoints used to have around 5-7 points hard-coded into it for testing purposed. This allows us to test much larger data sets, possibly independent of a bug that may be occurring when the number of points isn't a multiple of the workgroup size; with this function, we can set the number of points to be a multiple of the workgroup size, whereas the matrices have a set number of points whose number we can't control.