Naive: fix bug where # of points was not a multiple of thread group size
There was previously a bug where we may read out-of-bounds within the WebCL kernels if the number of points wasn't a multiple of the thread group size. The CL kernels look at the number of thread groups to determine the number of tiles, and the threads per thread groups to determine the size of the tiles. It then assumes each tile/thread group has the same number of points. However, if the number of points isn't a multiple of the size of the thread groups, then the last thread group would have fewer points/threads than all the rest.

This change makes it so that we explicitly check we're reading in-bounds.