Naive: use stepNumber to add randomness when there's overlapping points
In the WebCL kernels, when two points overlap, we apply a random force in place of the force of the other point (which is always 0) to separate the two points. We get the random force from the randValues array we pass into the kernel.

Previously, we chose the specific random value from the array based off the thread ID, which never changed for a point. Thus, it would have the same random value every time. Now, we multiply the stepNumber in to the calculation, so that the specific random value used changes on each step. This avoid problems where the random value doesn't create a strong enough force to 'split' the two overlapping points (since the next frame's random values may have such a property, and it only needs to happen once.)