Naive: fix bug where stepNumber wasn't getting set on modern WebCL
The polyfill version of setArg (used with old WebCL platforms) was updated to ignore null values, but the standard version (used with new WebCL platforms) wasn't. Because we rely on this behavior to set the kernel arg stepNumber, this value wasn't getting set on new WebCL platforms.

This fixes the behavior of setArg and, thus, stepNumber.