Naive: eliminate firstTick attribute from graph object
This was added before stepNumber was added to the graph. This was used to tell the graph to not run the simulator on the first tick, so that the first frame showed the graph points at their initial positions.

We now accomplish the same effect by simply checking the stepNumber.