Naive: remove dead code from GL shaders. Remove 3rd party copyright.
90% of the code in the GL shader has been commented out. This commit removes that code, as it's unlikely to be useful with our current architecture.

The original shaders were taken from the Samsung WebKit-WebCL project's WebCL example suite. Because of this, they carried a Samsung copyright notice and disclaimer. However, with the removal of dead code above, all that's left is the most basic remaining code, which is original work. Therefore, since all the original Samsung code is gone, we have removed the Samsung copyright notice.