Naive: when choosing random point, use real point, not made-up one
When two points overlap, their force = 0. We don't want these overlapping points to get 'stuck' together, so in the CL kernels, if we detect the two points are too close, we randomly choose another point, and use that point's forces in place of the overlapping point's. In this way, we 'nudge' two points apart.

Previously, calculatePointForce() had the responsibility of detecting if the points were overlapping. It would then create a new point entirely out of random values, and use that point instead.

This commit extracts the logic of choosing a new point into a new function, randomPoint(). Instead of creating a new random point, it randomly chooses one of the existing points to use instead. This should make the forces we're substituting more believable and regular. The caller of calculatePointForce() is now responsible for detecting overlapping points and calling randomPoint() to get a substitution; calculatePointForce() now blindly calculates the force of the two points it's given.

We have finally renamed calculatePointForce() to just pointForce().