Naive: allow RenderGL.createBuffer() to write data on creation
Previously, RenderGL.createBuffer() had an argument 'size', which you would think would set the size of the buffer we just created. But it didn't do anything. The buffer was only set to the right size when you wrote to it.

Now, RenderGL.createBuffer() takes an optional parameter, 'data', which will set the contents of that buffer to hold the given data. Alternatively, if you pass a number as 'data', the buffer will be set to that size, but no data written to it.