Naive: create {NBody,SimCL}.setEdges() function. Call it on main setup.
setEdges() will create the buffers necessary to hold the edge data, then write the given edges to it.

At the moment, the edge data is fake testing data (generated by NBody.setEdges(), which ignores the real edges passed to it.)

When main.js loads data (be it random starting data, or from a matrix load,) it now calls setEdges() in addition to setPoints(). However, in the case of random data, it passes in an empty edge list. (Though it doesn't make a difference since setEdges() is always loading test data atm.)