Use extend() wrapper rather than jquery directly
We create our own extend() function in the util module, rather than using jquery.extend() directly in RenderGL. Even though our own function is just currently a wrapper around jquery, this allows us to tweak extend()'s functionality in the future. It also means that we can use RenderJS without having to include all of jquery, if we decide to remove it in the future.

My editor also trimmed some trailing whitespace in RenderGL, but that should have no effect on functionality.