Reorder RenderGL.create() code to be clearer
This shouldn't change the code at all. We just reorder the code in RenderGL.create() to be clearer. Now, we do most of the work of creating and populating the renderer object at the beginning of the function, along with setting WebGL defaults and the like. The only stuff we do in the callback for createProgram() is stuff that involves programs.