In RenderGL, change where we find attr location
Instead of finding them as soon as we create the programs, and saving their values for later, we find their location when we set the attrib pointers for these attributes, in render().

This makes more logical sense, and is overall cleaner. It may cause a slight speed loss (since we ask GL for the attribute location every frame, instead of just when create the program,) but we can always cache this value, or something similar, if we need to fix it up later.