Trim trailing whitespace in NBody.js
We really should just do this in one go instead of piecemeal like this, but I'm lazy, and my editor automatically does this on save. So each time I save some file for the first time, it strips out the extra whitespace, and we end up with these commits.

The More You Know™