Remove 'kernel' stats graph
This was not giving accurate results because it was triggered for each of our multiple kernel calls. Therefore, it was more of an average of all kernels than the cumulative time it took for all kernels to execute per frame. It also seemed suspiciously low, even though the 'end' call was guarded by a queue.finish() call (which should theoretically block until the kernel finished executing.)