Move to new SimpleEvents.js-based event system
Instead of using the ad-hoc, brittle event system we were using, we now use an event system based off the SimpleEvents.js library I wrote.

This commit updates the event triggers and listeners to use this new system (which requires no added lines or complexity,) and removes a bunch of code required for the old system (which reduces complexity.) Yay!