Rename bound 'cl.copyBuffer()' to 'copyInto()'
In its normal form, copyBuffer() takes a source and destination argument, and so it's clear what it's doing. However, when it's bound to the source buffer, the source argument curried in. In this form, which is almost always how we actually call it, it takes only one argument: the destination buffer.

Because copyBuffer() with one argument is ambiguous (are we copying the bound buffer into the argument, or copying the argument into the bound buffer?), in its bound form, we renamed it copyInto().

This is the only instance where the bound function has a different name than the unbound function. In this case, I think it's the best solution.