Remove stepNumber arg from SimCL.set*()
SimCL.setPhysics() and SimCL.setLocked() both previously took the current stepNumber as an argument, and set it in the kernels they touched. However, this kernel argument is set every tick in the tick() function. This makes setting the arg anywhere else redundant.

We remove the stepNumber arg from both functions (and their callers.) In the calls to cl.setArgs() inside both function, we pass a null value where the stepNumber arg would go, so that setArgs() ignores that arg.