Remove unneeded buffer releases, & release sooner
In SimCL.tick(), the only buffers we acquire() are the two shared CL/GL buffers. However, at the end of tick(), we were calling release() on springsPos and midSpringPos as well. This was unneeded, and has been eliminated.

Further, instead of waiting until the end of tick() to release both buffers, we release curPoints as soon as the regular (non-midpoint) kernels finish, since the buffer isn't used by subsequent kernels.