`run-site`: launch WebKit.app on the same desktop
This commit makes it so that the run-site script will always launch WebKit.app in the same desktop every time.

Previously, when run-script re-launched WebKit.app after detecting it exited, it would often launch in the desktop of the last used app. This meant that, for example, WebKit would launch on top of Sublime Text, rather than re-launching at the same position it occupied when we first launched it (often on top of the Terminal window we ran the `run site` command from.)

This change makes it so that we record the position the WebKit window occupies when we first launch it. Then, on subsequent launches, we move the WebKit window back to that position (though we specifically retain the existing width and height, rather than setting them to the width and height of the previously recorded WebKit window.)

This should mostly work. However, it's possible that there are cases when it fails (if we move things around a lot, or changes desktops in a funny way, etc.) In that case, using CTRL+C to quit the script and re-running it should fix the problem.