Make it _fast_: use ArrayBuffer vs. Float32Array
For some reason, when sending binary data as a Float32Array (a subclass of TypedArrayView), socket.io would send the data as utf8 encoded binary data instead of raw binary data. It still was 'binary' data in the wire (looking at the Chrome network log panel, and the Wireshark packet capture,) but the client needed to do a string decode on the other end, which took FOREVER.

Sending a ArrayView object instead of a Float32Array gives 1-2 magnitudes of performance gain in sending the data.