Fix bug: we weren't deleting CL buffer properly
We were calling our wrapped buffer object's release() function, which was just a noop (if a regular CL buffer) or a enqueueReleaseGLObjects() (if a GL shared buffer.) This (poorly named) function is really meant to give up ownership of a buffer, not delete it.

We almost certainly meant to call the actual CL buffer's release() function, which deletes the CL buffer. This commit fixes it so that we call that function, rather than our wrapped release().