Ensure we're always releasing GL buf ownership
We now do gl/cl finish() calls when acquiring & releasing shared CL/GL buffers, and when we finish a simulator tick (we already did so after finishing a renderer tick.) This ensures that gl is done with the buffer before we acquire it, and CL is done with it before we render it.

The documentation for shared CL/GL objects says to call glFinish() to ensure that *CL* is done using the object. Not sure if this is a typo and they meant to say clFinish(). We call both clFinish() and glFinish() when CL is done with the buffers and wants to release them to GL, just to be sure.

This may slow things down with unnecessary blocking calls. We'll evaluate as we go.