In jshint config, remove global "io"
This was carried over from the client project's .jshintrc (where socket.io's client library is made available via a <script> tag on the HTML page, and thus jshint wasn't aware of its existence.) Here, we manually include the socket.io library as 'io', so a global define only interferes with that.

Also, we manually toggle that option off in server.js, just in case we're using a shared .jshintrc which still has the global option set.