Renderer: remove preserveDrawingBuffer in GL init
We added this so that the node-webgl would work, but WebGL in the browser works just fine without it, and the WebGL spec says that setting it to true may have "significant performance implications". So we let it stay and its default "false" value by removing it from the getContext() options.