Fix a bunch of bugs where cached values were wrong
In renderer.js, we created a number of caches for things like program bindings, attribute locations, etc. However, for a number of these, we uses WebGL objects as the keys. The problem is, two different WebGL objects of the same type will often/always resolve to the same string, and thus our caches were all messed up because they thought everything had the same key.

We now use strings containing object names in those places. This also means we have to pass in those names to a lot of functions. There's a better way to do this, but this works for the moment.