main.graph.js: wrap the render loop in try/catch
Previously, if an error was thrown inside this loop, it would be caught by socket.io (since it was thrown inside the event handler for "vbo-update" message) and it would be detected by our code as a connection error.

By wrapping it in a try/catch block, we can display a more accurate error message.