In dev mode, show "Disconnect" button
Helps when developing because we can disconnect the browser without shutting down the server. Often, we want to pause the browser client while we work, since it takes up tons of CPU, then reload it to see changes that we've made.

Previously, we'd either have to close the browser tab and re-open it when we wanted to run the client again, or shut down the server and re-run it when we wanted to run the client again. This latter mode was particularly bad because sometimes the server sim runs fast when opening, and sometimes it runs really slow, and it stays in either mode for the like of the process. No idea why some launches are different than others, but when we open it and it's running fast, we want to keep it running as long as possible.