Minor tweak to how we initialize CL/GL sharing
Changes the capitalization of the name of WebCL CL/GL sharing extension to match the spec exactly. It should not make a difference, but just in case...

Also (redundantly) enables the CL/GL sharing extension on the device itself, in addition to the previous behavior of enabling it on the context, in the node-webcl code path. Again, should not necessarily make a difference, but just in case...