Graph streamer: support for arbitrary hostname
Previously, StreamGL's client.js and nodecl's graph.html both assumed the server was always reachable at localhost. This commit modifies those files to use relative URLs where possible.

The one place this is not possible is in the client's XHR fetch of compressed VBOs (because the server is listening for XHR requests on a different port than it's serving the web page, and it's not possible to construct a relative URL with a different port.) In this case, client.js looks up the window.location of the page it's running on and uses that to construct an absolute URL to the same server, but a different port.

These changes collectively allow the server to be located at any arbitrary address (including those hosted on a different machine than the client,) and have everything still work.