Remove explicit benchmarking code from cl.js
There were places where we were recording the time at start and end of a task, to see how long it took, and writing it out in the debug messages. However, the `debug` module already prints relative time differences between messages from the same module, so this is redundant. This may also help speed up our code a tiny bit.