cl.js: eliminate polyfill() function
We now only support the latest WebKit WebCL and nodejs's node-webcl. When the API for these differs, cl.js tries to handle it with as minimal of changes as possible.

cl.create() (as of commit e3626f) has two different functions defined for node/the browser, and dynamically dispatches to one based on the environment. The WebCL APIs used by cl.call() and cl.setArgs() differed minimally, so we have one function with an if() statement inside it. Setting global vars (`webcl`, `types`, `DEVICE_TYPE`) is done by a simple if() statement at the end of the module, run on module initialization.