cl.js: createBufferGL uses createBuffer if no gl
In cl.createBufferGL(), if it is detected that the 'gl' buffer we're creating a CL buffer for is not, in fact, using GL, it calls cl.createBuffer() to create a regular CL buffer. If the non-gl (fake) 'buffer' has data associated with it, we also write it to the CL buffer before returning.