node-driver: use dummy renderer, and remove WebGL
This commit adds the RenderNull renderer class, which more-or-less has noop for all of its externally called functions. It should be a drop-in replacement for RenderGL for when we don't want to do rendering.

We also modify node-driver.js to use RenderNull instead of RenderGL. All use of node-webgl has been removed (we no longer create a window/document, canvas, etc.) node-driver just runs the simulator event loop, and server.js uses it to server it to web clients.