server: reorder middleware so static served last
The order of express middleware in server.js is changed so that express.static is last, so that static files are only served if none of the other routes/middleware can respond to the request.

The call to http.listen() is also moved to the end of the file so that we only start listening for requests after all routes/middleware are setup.