Fix data-loader.js when no matrix data available
We've removed the matrix data from this repo (since it's huge and we
aren't currently using it,) but the data-loader.js function that gives
a list of all available data still tried to read
`data/matrices.binary.json`, which caused an error. Now it first checks
that this file exists, then loads it if available, or returns an empty
array if not.