(Prototype) separate 'master' and 'viz' servers
This commit adds prototype-quality (read: not production ready) support
for running one server serving static content + routing clients, and
another serving viz data.

To run this, first make sure you've checked out the branch of the same
name in the Superconductor2 repo and `npm install` + `grunt` in the
StreamGL directory. Then run `server.js` and `server-viz.js` in this
repo and open your browser to localhost:10000. The graph demo should
work as normal, but the data is being served from 2 different server
processes.

At the moment, the hostname and port of the viz server is hard-coded to
be 'localhost' and (the configured server address + 1), respectively.
This means that it will NOT work on the server (or running on anything
other than localhost.) This can be changed at the top of server.js +
changing the listen port at the bottom of server-viz.js. All users also
get the same hostname and port.