Remove local files (can revert later if needed)
This commit removes all the files used for running the graph viz in
local (i.e. non streaming ec2 server mode). This module is currently
primarily intended for streaming server use, so we remove these files
from the repo.

If we decide we want this module to support local mode later, we can
recover some/all of the files from Git history.