Split client.init() & use server render-config
Instead of having one monolithic function which connects to the
websocket server, initializes the renderer, and handles the VBO update
messages, we split it into three functions, one for each of the
mentioned tasks. Each returns an Rx observable that fires when it is
done, so that they can be chained.

Further, the createRenderer() function also fetches the render-config
from the server, rather than using it via a module include. This allows
us to get rid of the hacky <script> tag stuff we were doing to assign
different render-configs to different frontends.