Revert StreamGL dep in package.json to point to GitHub
An earlier commit had changed the dependency to be just "*". This would
only work if we did a `npm link StreamGL` before installing (or
StreamGL was installed globally.)

This reverts that dependency to again point to the GitHub repo hosting
StreamGL. Installing graph-viz should now work without linking StreamGL
first.

However, if you *do* want to use a locally-installed StreamGL copy
instead of downloading StreamGL from GitHub, just perform a `npm link
StreamGL` in the graph-viz directory, and it will use the local version
from then on (even with the dep pointing to GitHub.)