Duplicate some dependencies to devdependencies as a reference
We previously moved a bunch of devdependencies to the regular
dependencies section of package.json, since we need the full
dependencies to build the StreamGL.js file on install. We copy these
dependencies back to devdependencies (where they are redundant with
dependencies now) as a reference of what dependencies we can remove
later when this module no longer needs to be self-building.