iPad: Ignore swipes while pinching. Fixes window blanking during pinch.
The base problem for the window blanking was that that, if you're
currently pinching (zooming in/out) and also happen to do a drag (which
is easy to accidentally do,) the drag event object will have an array
of positions as its current position, rather than the expected single
position x,y. We suspect this is because Quo reports a position for
each finger on the screen (which is two during pinches.)

This commit changes the drag even handler to ignore drag events that
have more than one position reported.

We tried just using the first reported position as the position for the
drag event, but we noticed lots of jumpiness in panning/zooming when we
did so. It's probably correct to ignore the drag event during pinches,
but there might be a more elegant way to do so than we're doing here
(something with Rx stream?). But this works for now.