Use our fork of bootstrap-slider to get bugfix
There is a bug in mainline bootstrap-slider where it will fail to
install as a npm dependency. I've submitted a pull request at
https://github.com/seiyria/bootstrap-slider/pull/255 . Until that is
merged into master and a new version published to npm, we should
*temporarily* move the dependency listing to point to our fork on
GitHub.

Once the pull request is merged and published on npm, we should move
back to the standard bootstrap-slider dependency.