Set graph-viz sliders to use linear scale
Previously, while the sliders themselves showed a linear change in
values, the values sent to the server were exponential. This changes it
so that the values you see in the browser match the changes seen in the
real values (times a constant coefficient of 1000 so that the values
seen in the browser aren't all small decimals).