Update slider bounds to prevent viz explosion
There's a bug in the simulator where certain extreme values for the
physics constants cause the visualization to start jumping and
jittering, with geometry eventually jumping off the screen, never to
return, and the viz crashes.

We update the bounds of the HTML sliders in the demo to try to prevent
these extreme values. This doesn't completely prevent the problem (as
it's combination of values that cause it, so values that are otherwise
benign in one constant can cause an explosion when another constant is
set to certain values) but it helps.

We also update the slider step value to try to give 100 distinct values
within the new bounds.