Send initial slider values to server on load
Previously, the server started with some default values for the physics
constants in the graph simulation. The HTML sliders on the demo page
also showed some default values for the physics constants. They didn't
necessarily match. This meant if you moved the slider just a bit, the
HTML value would be sent to the server and updated in the simulation,
and this was potentially very different from the previous server value.
This caused the simulation to change suddenly, rather than with the
small increment the user had performed.

This commit changes it so that the initial values of the HTML sliders
are sent to the server on page load, meaning that the sim *should*
start with those values.

There is a potential bug where we send all these settings as
independent messages on page load. We should verify that getting these
4 messages all at once won't cause any performance problems with the
graph-viz server.