Update bootstrap-slider dep to use their repo
Per commit 73f04b7faa9fe, we changed the bootstrap-slider dependency in
package.json to point to our own fork of the official repo, in order to
fix a bug. As of this commit, that bugfix has been merged back into the
main bootstrap-slider GitHub repo, but not yet published as a new
version on npm. We therefore update our package.json to point to the
official GitHub repo directly.

Once they publish a new version of bootstrap-slider on npm, we should
use that version.