Fix browserify bug by making labeled-stream-splicer@1.0.0 a direct dep
browserify has labeled-stream-splicer as a dependency with the version
constraint of '^1.0.0'. That module just published v1.0.1 and it broke
things. Now browserify gets a "module not found" error when doing a
`require('labeled-stream-splicer')`.

Since 1.0.1 satisfies the version constraint of browserify, it will
continue to use that broken version by default. However, this commit
sets labeled-stream-splicer v1.0.0 (exact version) as a dependency of
StreamGL. npm will install labeled-stream-splicer v1.0.0 first, see
that it satisfies the browerify dep and use that version, and
everything will continue to work.

npm sucks.