Split MONGO_SERVER option into parts (but leave MONGO_SERVER for now)
Instead of crafting one giant connection string for Mongo in the
MONGO_SERVER option (which doesn't give us insight into the details of
the connection, or granularity to change it,) this commit splits it
into MONGO_USERNAME, MONGO_PASSWORD, MONGO_HOSTS, and MONGO_DATABASE. A
complete connection string can then be constructed from these options
from any code that needs to connect to Mongo.

The DATABASE option is already redundant with the MONGO_SERVER option,
so that should also be removed in the future.

The old 'MONGO_SERVER' and 'DATABASE' options remain for now, for
backward compatibility reasons.