Remove explicit IP addresses for servers in inventory; use only DNS
We used to give each server in the inventory an explicit IP address to
use instead of its DNS name. So, for example, we had
`staging.graphistry.com ansible_ssh_host=54.183.155.35`, which meant
Ansible would actually connect to 54.183.155.35 when told to connect to
staging.graphistry.com.

This allowed us to connect and configure machines even when they didn't
have their DNS name set to point to them yet, or when we were doing a
hot-swap and the DNS pointed to the old server, but we wanted to
configure a new server to take its place (since the domain name is what
we configure the server's hostname as, it was important we used the
eventual final DNS name rather than its current IP.) However, we
allowed these IPs to get out of sync with the actual hostnames, and the
"production" inventory in particular had the wrong IPs for the hosts it
was trying to connect to.

We remove these IPs so that we can always connect to the right system.
If we have an edge-case where the systems don't have the correct DNS
record yet, we can go back and set the ansible_ssh_host property of
those hosts.