Splunkfwd: change group of deploymentclient.conf: root->graphistry-dev
We may have not run the users role (which creates the graphistry-dev
group) at the point where we install the splunk forwarder, so we
shouldn't set the owner of the Splunk deploymentclient.conf file (which
configures the splunk indexer as a deployment client of the Splunk
deployment server) to graphistry-dev, or we might error. Instead, use
the root group, which is what most splunk local config files are owned
by normally anyway.