Worker: change default # of viz worker processes to 8, set prod to 16
We reset the default number of worker processes to 8 (after increasing
it to 32 in our last commit.) This is the value that will be used on
staging.

However, for production, we override this value to be 16.

The reason for two values (as I'm told) is that it will allow us to
discover bugs more easily on staging (since zombie processes are more
likely to cause out-of-worker errors more quickly.) On production, it's
better to hide bugs as much as possible, so it has a higher value.
Plus, it allows us to support more simultaneous users. We only use 16
(vs. a higher value) because the previously deployed value was 8, and
we should probably change this value in a smaller step so that we can
see what effect it has.