Memoize the exported function to get options
We import the config module and call the function it exports (which
returns the current config options and values) from a ton of places,
often multiple times in the same application (since each file and/or
submodule that needs configuration values does this.) Previously, we'd
re-read every configuration source and combine them together to resolve
the current options set each time this function was called (logging the
config with the `debug` module when we did.) This was a lot of
redundant work.

This commit changes that function to be memoized (using lodash'
built-in 'memoize function' feature.) Since the only argument to the
`getOptions()` function we export is any option overrides we want to
apply, we make the memoization key the JSON string form of the first
argument.