nodejs: change installed `npm` version from `latest` to `2.7.3`
There is a bug in `npm@2.7.4` (the current `latest` as of writing this,
and thus the default installed) that causes errors when trying to
install a module from a private GitHub repo. There is a further bug in
`npm@2.7.5` that causes errors messages to be thrown under the same
conditions; the module seems to install anyway, but the STDERR output
is enough for Ansible to think the install failed and stop running the
playbook.

We therefore explicitly install `npm@2.7.3` (the last non-buggy version
as of April 1st, 2015) rather than `latest`. We further add a warning
to the Ansible about this, and advise un-doing this change as soon as
`npm` works again.