Users: don't upload ~/.ssh/config file, and delete existing copies
I'm not sure if this change (from recent commit 0f7508c5e0dd2) actually
helped the problem it was intended to address (SSH keys not getting
forwarded, causing git clones of GitHub modules to fail). I've come to
believe this problem actually stems from Ansible v1.9.0, since I only
see it when running Ansible on the deploy server (which has that
version), and not when running from my laptop (which has Ansible
1.8.4). Ansible 1.9.0 broke some other `sudo` and user/permissions
stuff, so it's my bet on the real cause of the bug.

Just in case I'm wrong, the upload task has only been commented-out,
not deleted, and the SSH config file we uploaded is still in this repo,
should we want to re-enable it.