Temp. bug fix: 'too many users' error, even when workers available
We were seeing a bug in the routing portion of central where it would report a "too many users..." error to the user, even when that was not true. This error is reported when central can't find any available workers to route the user to. Central was throwing that error even when there were a ton of available workers, which was a bug.

The cause was that when attempting to route a user, central could potentially handshake *all* workers. A handshake reserves the worker it was applied to for `n` seconds, and so if a second user connected to central within `n` seconds and we tried to route them, all workers would report that they're busy when we tried to handshake them, and the "too many users" error would be thrown.

Central only intended to handshake one worker at a time, continuing only until the first successful handshake. This was implemented as an Rx stream similar to `ips.flatMap(fn(ip){return handshake(ip);}).filter(failedHandshakes).first()`. The `first()` operator was intended to stop the handshake attempts after the first success. The problem is that Rx is fundamentally time-based, not item based.

So the `ips` Observable would emit all the IPs at once, and we would start firing (asynchronous) handshake requests to each. Only when the first successful handshake had **returned** would Rx dispose of the subscription, ending the handshake attempts. In other words, there was a race where Rx would fire off as many handshakes as it could before one returned success.

This commit fixes that bug by making `ips` a controlled Observable, i.e. one that only emits an item when we explicitly request it. We request the first one, fire off a handshake, and then wait for its result. If it's a failure, we request another IP; if it's a success, the IP is emitted, and `first()` will then end the stream.